requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
databricks-sdk>=0.18.0
databricks-sql-connector>=3.0.0
python-dotenv>=1.0.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0

